Move test to save core ### What's done: * Move test to save core
